refactor(runtime): split class GC-root scanner into class_gc_roots.rs
Move scan_class_inheritance_roots_mut + its cfg(test) seeds out of
class_registry.rs into a sibling object/class_gc_roots.rs and re-export from
object/mod.rs, keeping class_registry.rs under the 2,000-line CI gate. No
behavior change.
